Order the cli
