+++ backoffice_backend/src/utils/cliHandlers/Handlers.ts
@@ -1,0 +1,8 @@
+import createUser from './createUser';
+
+export const handlers = {
+  createUser: {
+    handler: createUser,
+    allowedValues: ['username', 'password', 'email'],
+  },
+};
