+++ backoffice_backend/src/utils/cliHandlers/typesAndConsts.ts
@@ -1,0 +1,10 @@
+export const ArgsRegex = /--(.*)=(.*)/gm;
+
+export type HandlerOptions = {
+  [key: string]: any
+}
+
+export type ActionNameAndOptions = {
+  action: string,
+  options: HandlerOptions
+};
